chore: clean up project structure, move mock data, and update imports
- Move mock data (features, stats, steps) from types/data to mock/data
- Remove unused/duplicate files in types/data and components root
- Update all imports to use new mock/data location
- Clean up and organize components and types folders
- Fix map error in HeroSection and image path issues
- General structure and code hygiene improvements
