+++ mock/data/features.tsx
@@ -1,0 +1,34 @@
+import { Shield, Eye, TrendingUp, Zap, Globe, Users } from "lucide-react"
+
+export const features = [
+  {
+    icon: <Shield className="w-6 h-6" />,
+    title: "Blockchain Security",
+    description: "Immutable transactions with smart contract automation for guaranteed security and compliance",
+  },
+  {
+    icon: <Eye className="w-6 h-6" />,
+    title: "Real-Time Tracking",
+    description: "Track every donation from source to impact with complete transparency and audit trails",
+  },
+  {
+    icon: <TrendingUp className="w-6 h-6" />,
+    title: "NGO Scoring",
+    description: "Dynamic performance evaluation based on fund utilization, compliance, and project outcomes",
+  },
+  {
+    icon: <Zap className="w-6 h-6" />,
+    title: "Less Fees",
+    description: "Direct peer-to-peer transactions eliminate middlemen and reduce costs to near zero",
+  },
+  {
+    icon: <Globe className="w-6 h-6" />,
+    title: "Cryptocurrency",
+    description: "Accept donations in multiple cryptocurrencies for fast, borderless transactions",
+  },
+  {
+    icon: <Users className="w-6 h-6" />,
+    title: "Privacy Options",
+    description: "Choose to donate anonymously or publicly while maintaining complete transaction transparency",
+  },
+]
