+++ mock/data/stats.tsx
@@ -1,0 +1,6 @@
+export const stats = [
+  { value: "14 ETH", label: "Total Donations" },
+  { value: "3", label: "Verified NGOs" },
+  { value: "15", label: "Active Donors" },
+  { value: "100%", label: "Transparency Rate" },
+]
