+++ types/data/stats.tsx
@@ -1,6 +1,1 @@
-export const stats = [
-  { value: "₱ 1.2M", label: "Total Donations" },
-  { value: "69", label: "Verified NGOs" },
-  { value: "1,000+", label: "Active Donors" },
-  { value: "99.9%", label: "Transparency Rate" },
-]
+
