+++ types/data/steps.tsx
@@ -1,17 +1,1 @@
-export const steps = [
-  {
-    number: "01",
-    title: "Connect & Browse",
-    description: "Sign in with Web3Auth and discover verified NGOs with performance scores",
-  },
-  {
-    number: "02",
-    title: "Smart Contracts",
-    description: "Funds are secured in smart contracts and released based on compliance requirements",
-  },
-  {
-    number: "03",
-    title: "Track Impact",
-    description: "Monitor real-time fund utilization and project outcomes through blockchain records",
-  },
-]
+
